+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   service: `--no-rotation` to disable the internal secret rotation process and
   `--no-authn-local` to disable the internal local authentication socket server.
   CNJR-3503
+- Adds support for optionally prefixing user role_id with "user/" during API key authentication.
+  CNJR-5214
 
 ### Fixed
 - Dedicated user identifier resolver allowing the user identifiers work like any other resource id. The Conjur internal